refactor: move leftover e2e_support pin hosts into test/support (#4598)
Keep widget-test MaterialApp/pump hosts out of pin suites so the AST
gate can forbid `_pumpHost` clones without changing helper timings.
