+++ packages/colonizethis_app_e2e_support/test/e2e_open_civilian_panel_test.dart
@@ -28,15 +28,13 @@
 /// timeout regressions).
 library;
 
-import 'dart:async';
-
 import 'package:colonizethis_app_fixtures/config/ct_e2e.dart';
 import 'package:colonizethis_app/features/game/screens/game/game_screen_shared.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
-import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:colonizethis_app_e2e_support/e2e_test_shared.dart';
+import 'support/e2e_open_civilian_panel_hosts.dart';
 import 'support/e2e_widget_pump_harness.dart';
 
 void main() {
@@ -45,7 +43,7 @@ void main() {
   testWidgets(
     'e2eOpenCivilianPanel taps the empire rail button and detects the panel',
     (WidgetTester tester) async {
-      await tester.pumpWidget(wrapE2eApp(_CivilianRailHarness()));
+      await tester.pumpWidget(wrapE2eApp(CivilianRailHost()));
       expect(find.byKey(kEmpireCivilianUnitsButtonKey), findsOneWidget);
       expect(find.byKey(kCtE2ECivilianPanelRootKey), findsNothing);
       await e2eOpenCivilianPanel(tester, timeout: const Duration(seconds: 5));
@@ -62,7 +60,7 @@ void main() {
 
   testWidgets('e2eOpenCivilianPanel falls back to the first-civilian marker '
       'when the empire rail button is absent', (WidgetTester tester) async {
-    await tester.pumpWidget(wrapE2eApp(_CivilianMarkerOnlyHarness()));
+    await tester.pumpWidget(wrapE2eApp(CivilianMarkerOnlyHost()));
     expect(find.byKey(kEmpireCivilianUnitsButtonKey), findsNothing);
     expect(find.byKey(kCtE2EOpenFirstCivilianMarkerPanelKey), findsOneWidget);
     expect(find.byKey(kCtE2ECivilianPanelRootKey), findsNothing);
@@ -83,7 +81,7 @@ void main() {
     (WidgetTester tester) async {
       await tester.pumpWidget(
         wrapE2eApp(
-          _DelayedCivilianPanelHarness(mountAfter: Duration(milliseconds: 120)),
+          DelayedCivilianPanelHost(mountAfter: Duration(milliseconds: 120)),
         ),
       );
       expect(find.byKey(kCtE2ECivilianPanelRootKey), findsNothing);
@@ -123,128 +121,4 @@ void main() {
       );
     },
   );
-}
-
-/// Test harness that mounts the civilian panel root synchronously when the
-/// keyed empire rail button is tapped, so the helper detects the panel on
-/// the very next frame.
-class _CivilianRailHarness extends StatefulWidget {
-  const _CivilianRailHarness();
-
-  @override
-  State<_CivilianRailHarness> createState() => _CivilianRailHarnessState();
-}
-
-class _CivilianRailHarnessState extends State<_CivilianRailHarness> {
-  bool _panelOpen = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Column(
-        children: <Widget>[
-          TextButton(
-            key: kEmpireCivilianUnitsButtonKey,
-            onPressed: () => setState(() => _panelOpen = true),
-            child: const Text('Civilian'),
-          ),
-          if (_panelOpen)
-            const KeyedSubtree(
-              key: kCtE2ECivilianPanelRootKey,
-              child: SizedBox(width: 100, height: 100),
-            ),
-        ],
-      ),
-    );
-  }
-}
-
-/// Test harness that exposes the first-civilian marker but no empire rail
-/// button, so the helper must take the marker branch as the second canonical
-/// trigger.
-class _CivilianMarkerOnlyHarness extends StatefulWidget {
-  const _CivilianMarkerOnlyHarness();
-
-  @override
-  State<_CivilianMarkerOnlyHarness> createState() =>
-      _CivilianMarkerOnlyHarnessState();
-}
-
-class _CivilianMarkerOnlyHarnessState
-    extends State<_CivilianMarkerOnlyHarness> {
-  bool _panelOpen = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Column(
-        children: <Widget>[
-          TextButton(
-            key: kCtE2EOpenFirstCivilianMarkerPanelKey,
-            onPressed: () => setState(() => _panelOpen = true),
-            child: const Text('Marker'),
-          ),
-          if (_panelOpen)
-            const KeyedSubtree(
-              key: kCtE2ECivilianPanelRootKey,
-              child: SizedBox(width: 100, height: 100),
-            ),
-        ],
-      ),
-    );
-  }
-}
-
-/// Test harness that mounts the civilian panel root only after an in-test
-/// timer fires so the helper exercises its adaptive post-tap waits before
-/// returning. The keyed empire rail button drives the schedule, mirroring
-/// the production-panel async harness pattern in
-/// `app/test/e2e_open_production_panel_test.dart`.
-class _DelayedCivilianPanelHarness extends StatefulWidget {
-  const _DelayedCivilianPanelHarness({required this.mountAfter});
-
-  final Duration mountAfter;
-
-  @override
-  State<_DelayedCivilianPanelHarness> createState() =>
-      _DelayedCivilianPanelHarnessState();
-}
-
-class _DelayedCivilianPanelHarnessState
-    extends State<_DelayedCivilianPanelHarness> {
-  bool _panelOpen = false;
-  bool _scheduled = false;
-
-  void _handleTap() {
-    if (_scheduled) {
-      return;
-    }
-    _scheduled = true;
-    Timer(widget.mountAfter, () {
-      if (!mounted) {
-        return;
-      }
-      setState(() => _panelOpen = true);
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Column(
-        children: <Widget>[
-          TextButton(
-            key: kEmpireCivilianUnitsButtonKey,
-            onPressed: _handleTap,
-            child: const Text('Civilian'),
-          ),
-          if (_panelOpen)
-            const KeyedSubtree(
-              key: kCtE2ECivilianPanelRootKey,
-              child: SizedBox(width: 100, height: 100),
-            ),
-        ],
-      ),
-    );
-  }
 }
